fix(cluster): gracefully stop worker threads

+++ packages/cluster/src/utils/mode/impl/worker_threads/app.ts
@@ -1,6 +1,8 @@
+import { once } from 'node:events';
 import { setTimeout as sleep } from 'node:timers/promises';
 import { Worker as ThreadWorker, type WorkerOptions } from 'node:worker_threads';
 
+import { WORKER_THREAD_GRACEFUL_EXIT } from '../../../../worker_protocol/worker-thread.ts';
 import type { MessageBody } from '../../../messenger.ts';
 import { BaseAppWorker, BaseAppUtils } from '../../base/app.ts';
 
@@ -139,12 +141,22 @@ export class AppThreadUtils extends BaseAppUtils {
     return this;
   }
 
-  async kill(): Promise<void> {
-    for (const worker of this.#workers) {
-      const id = Reflect.get(worker, 'id');
-      this.log(`[master] kill app worker#${id} (worker_threads) by worker.terminate()`);
-      worker.removeAllListeners();
-      worker.terminate();
-    }
+  async kill(timeout: number): Promise<void> {
+    await Promise.all(
+      this.#workers.map(async (worker) => {
+        const id = Reflect.get(worker, 'id');
+        this.log(`[master] gracefully close app worker#${id} (worker_threads)`);
+        worker.removeAllListeners();
+        const exited = once(worker, 'exit').then(
+          () => true,
+          () => false,
+        );
+        worker.postMessage(WORKER_THREAD_GRACEFUL_EXIT);
+        if (!(await Promise.race([exited, sleep(timeout).then(() => false)]))) {
+          this.log(`[master] terminate app worker#${id} after ${timeout}ms timeout`);
+          await worker.terminate();
+        }
+      }),
+    );
   }
 }
